move relationship class methods into the single reopen directly
